Move hpp files to h
