+++ cpp/src/arrow/util/alp/AlpConstants.h
