+++ cpp/src/parquet/decoder.cc
@@ -41,9 +41,9 @@
 #include "arrow/util/bit_util.h"
 #include "arrow/util/bitmap_ops.h"
 #include "arrow/util/byte_stream_split_internal.h"
-#include "arrow/util/alp/Alp.hpp"
-#include "arrow/util/alp/AlpConstants.hpp"
-#include "arrow/util/alp/AlpWrapper.hpp"
+#include "arrow/util/alp/Alp.h"
+#include "arrow/util/alp/AlpConstants.h"
+#include "arrow/util/alp/AlpWrapper.h"
 #include "arrow/util/checked_cast.h"
 #include "arrow/util/int_util_overflow.h"
 #include "arrow/util/logging_internal.h"
